bpf: fix iptables masquerading for node -> remote pod traffic
When Cilium runs with KPR, host-firewall or bandwidth manager, it will
try to auto-derive one or more devices to which the bpf_host program is
attached.

This program will, among other things, redirect ingress traffic destined
to a pod to the pod's lxc device using `bpf_redirect()`.

This causes the traffic to bypass the nf_conntrack table, leading to a
situation where traffic leaving the pod after the connection's been
established will be (incorrectly) masqueraded in case Iptables
masquerading is enabled, since the connection is not tracked by
netfilter.

This commit fixes this by skipping `bpf_redirect()` when we detect this
case (i.e. traffic is flowing through bpf_host attached to a physical
device and Cilium has installed Iptables rules which require conntrack).

Fixes: #14859
Suggested-by: Martynas Pumputis <m@lambda.lt>
Signed-off-by: Gilberto Bertin <gilberto@isovalent.com>